// frontend/src/lib/api/generated.ts
export interface components {
    schemas: {
// ...
        ConsentAcceptance: {
            /** @enum {string} */
            documentType: "age_18" | "user_terms" | "personal_data" | "public_profile_distribution";
            documentVersion: string;
            /** @constant */
            accepted: true;
        };
// ...
        ProfileInput: {
            fullName: string;
            specialization: string;
            timezone: components["schemas"]["TimeZone"];
            institute?: string;
            course?: number;
            department?: string;
            company?: string;
            position?: string;
            avatarMediaId?: components["schemas"]["Uuid"];
        };
// ...
    };
// ...
}